agentic_api/requirements.txt
langchain
langchain-community
langgraph
langchain-chroma
langchain-groq
groq
python-dotenv
fastembed
PyPDF2
pycryptodome
streamlit
streamlit-pdf-viewer
langchain-huggingface
streamlit-chat-widget
fastapi[standard]